Add Assign material operators
There's different operators for Edit mode and Object Mode.
The Assign material operator in Object mode lets the user choose to 
override/clear out (default) the existing material slots (if more than 
one), otherwise one will be appended. If there's just one material slot 
for the object, that will be overridden no matter the state of the 
"Override setting".
This keeps the material slots clean, but the user has an option to keep 
existing material slots.
(They both call the same function)
This of course does not work in edit mode, where we assign to different 
polygons rather than the object itself).

Thoughts:
We could change the materials of all of the slots instead (keeping their 
assignments).
If all polygons are selected in edit mode, we could override/clear out 
all the material slots