Add Material Slot Move
Add Material slot move up/down & to top/bottom to the Material context 
menu.
This is all now done with one operator (instead of four/one for each 
"movement").
The plus is neater/less repetitive code, the drawback is that we can't 
disable the menu options depending on the active materials position!