Fix Issue #4 (Assigning materials when slot is linked to object)
It now works OK. It's not perfect, but mainly because how Blender 
handles material (slots) on linked objects.
An effort has been made to "smooth it out", but it might not be the best 
way to handle it.

(Functionality has also been added that could let the user choose how 
the material should be linked on assignment)